chatbox: close current input before attempting to open a new one
We need to make sure the eventbus unregisters the current panel before opening a new one. Otherwise, singleton panels can be stuck in a state where they don't receive events if a fresh instance is supposed to be opened.

We also need to make sure to not identify the current input too soon because if a panel is created through scripts, MESSAGE_LAYER_CLOSE will fire and destroy the current input.